Separate out ws.rs
