+++ config.toml
@@ -4,7 +4,7 @@ accessible_host = "192.168.1.10"
 # op_mode = "proxy"
 op_mode = "native"
 # server_url = "http://192.168.1.10:8080"
-server_url = "https://ecast.jackboxgames.com"
+# server_url = "https://ecast.jackboxgames.com"
 
 [tls]
 cert = "certs/192.168.1.10.pem"
