+++ src/ws.rs
@@ -1,11 +1,9 @@
 use std::{
     borrow::Cow,
-    fmt::Display,
     io::{Read, Write},
     process::Stdio,
-    str::FromStr,
     sync::{
-        atomic::{AtomicBool, AtomicI64, AtomicU64},
+        atomic::{AtomicI64, AtomicU64},
         Arc,
     },
 };
@@ -16,16 +14,16 @@ use axum::extract::{
 };
 use dashmap::DashMap;
 use futures_util::{stream::SplitSink, SinkExt, StreamExt};
-use hex_color::HexColor;
-use serde::{de::Error, ser::SerializeMap, Deserialize, Serialize};
+use serde::{ser::SerializeMap, Deserialize, Serialize};
 use serde_json::json;
-use tokio::{
-    io::Interest,
-    sync::{Mutex, Notify},
-};
+use tokio::sync::{Mutex, Notify};
 use tokio_tungstenite::tungstenite::client::IntoClientRequest;
 
-use crate::{JBRoom, Role, Token, WSQuery};
+use crate::{
+    acl::Acl,
+    entity::{JBAttributes, JBDoodle, JBEntity, JBLine, JBObject, JBRestrictions, JBType, JBValue},
+    JBRoom, Role, Token, WSQuery,
+};
 
 type Connections = DashMap<i64, Arc<Client>>;
 
@@ -73,211 +71,6 @@ pub struct Room {
     pub exit: Notify,
 }
 
-#[derive(Serialize, Debug)]
-pub struct JBEntity(JBType, JBObject, JBAttributes);
-
-#[derive(Serialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct JBObject {
-    key: String,
-    val: JBValue,
-    restrictions: JBRestrictions,
-    version: u32,
-    from: AtomicI64,
-}
-
-#[derive(Serialize, Debug, Clone)]
-#[serde(untagged)]
-enum JBValue {
-    Text(String),
-    Number(f64),
-    Object(serde_json::Map<String, serde_json::Value>),
-    Doodle(JBDoodle),
-}
-
-#[derive(Serialize, Debug, Clone, Copy)]
-#[serde(rename_all = "lowercase")]
-enum JBType {
-    Text,
-    Number,
-    Object,
-    Doodle,
-}
-
-impl FromStr for JBType {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "text" => Ok(Self::Text),
-            "number" => Ok(Self::Number),
-            "object" => Ok(Self::Object),
-            "doodle" => Ok(Self::Doodle),
-            _ => Err(()),
-        }
-    }
-}
-
-#[derive(Deserialize, Serialize, Debug)]
-struct JBRestrictions {
-    #[serde(default)]
-    #[serde(skip_serializing_if = "String::is_empty")]
-    #[serde(rename = "type")]
-    data_type: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    min: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    max: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    increment: Option<f64>,
-}
-
-#[derive(Deserialize, Serialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct JBAttributes {
-    locked: AtomicBool,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    acl: Vec<Acl>,
-}
-
-impl JBAttributes {
-    fn perms(&self, role: Role, id: i64) -> Option<Interest> {
-        if role == Role::Host {
-            return Some(Interest::READABLE | Interest::WRITABLE);
-        }
-
-        let mut perms: Option<Interest> = None;
-
-        for principle in self.acl.iter() {
-            if principle.principle.matches(role, id) {
-                *perms.get_or_insert(principle.interest) |= principle.interest;
-            }
-        }
-
-        perms
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-struct Acl {
-    interest: Interest,
-    principle: Principle,
-}
-
-impl Acl {
-    fn default_vec() -> Vec<Self> {
-        vec![Self::default()]
-    }
-}
-
-impl Default for Acl {
-    fn default() -> Self {
-        Acl {
-            interest: Interest::READABLE,
-            principle: Principle::Wild,
-        }
-    }
-}
-
-impl Display for Acl {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        if self.interest.is_readable() {
-            write!(f, "r")?;
-        }
-        if self.interest.is_writable() {
-            write!(f, "w")?;
-        }
-
-        write!(f, " {}", self.principle)
-    }
-}
-
-impl Serialize for Acl {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        format!("{}", self).serialize(serializer)
-    }
-}
-
-impl<'de> Deserialize<'de> for Acl {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        let acl: Cow<'de, str> = Deserialize::deserialize(deserializer)?;
-        Acl::from_str(acl.as_ref()).map_err(|_| {
-            D::Error::invalid_value(
-                serde::de::Unexpected::Str(acl.as_ref()),
-                &"A valid ACL principle",
-            )
-        })
-    }
-}
-
-impl FromStr for Acl {
-    type Err = ();
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let split = s.split_once(' ').ok_or(())?;
-
-        let interest = match split.0 {
-            "r" => Interest::READABLE,
-            "w" => Interest::WRITABLE,
-            "rw" => Interest::READABLE | Interest::WRITABLE,
-            _ => return Err(()),
-        };
-
-        if split.1 == "*" {
-            return Ok(Self {
-                interest,
-                principle: Principle::Wild,
-            });
-        }
-
-        let split = split.1.split_once(':').ok_or(())?;
-
-        let principle = match split.0 {
-            "role" => Principle::Role(split.1.parse()?),
-            "id" => Principle::Id(split.1.parse().or(Err(()))?),
-            _ => return Err(()),
-        };
-
-        Ok(Self {
-            interest,
-            principle,
-        })
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-enum Principle {
-    Role(Role),
-    Id(i64),
-    Wild,
-}
-
-impl Principle {
-    fn matches(&self, role: Role, id: i64) -> bool {
-        match *self {
-            Self::Role(r) => role == r,
-            Self::Id(i) => id == i,
-            Self::Wild => true,
-        }
-    }
-}
-
-impl Display for Principle {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Role(r) => write!(f, "role:{}", r),
-            Self::Id(id) => write!(f, "id:{}", id),
-            Self::Wild => write!(f, "*"),
-        }
-    }
-}
-
 #[derive(Deserialize, Serialize, Debug)]
 #[serde(rename_all = "camelCase")]
 struct JBProfile {
@@ -318,55 +111,12 @@ struct JBCreateParams {
     doodle: JBDoodle,
 }
 
-#[derive(Deserialize, Serialize, Debug, Clone)]
-#[serde(rename_all = "camelCase")]
-struct JBDoodle {
-    #[serde(default)]
-    colors: Vec<HexColor>,
-    #[serde(default)]
-    live: bool,
-    #[serde(default)]
-    max_points: u32,
-    #[serde(default)]
-    max_layer: u32,
-    #[serde(default)]
-    size: JBSize,
-    #[serde(default)]
-    weights: Vec<u32>,
-    #[serde(default)]
-    lines: Vec<JBLine>,
-}
-
-#[derive(Deserialize, Serialize, Debug, Clone)]
-#[serde(rename_all = "camelCase")]
-struct JBLine {
-    color: HexColor,
-    weight: u32,
-    layer: u32,
-    points: Vec<JBPoint>,
-    #[serde(default)]
-    index: usize,
-}
-
 #[derive(Deserialize, Serialize, Debug)]
 #[serde(rename_all = "camelCase")]
 struct JBKeyWithLine {
     key: String,
     #[serde(flatten)]
     line: JBLine,
-}
-
-#[derive(Deserialize, Serialize, Debug, Clone, Copy)]
-struct JBPoint {
-    x: f64,
-    y: f64,
-}
-
-#[derive(Deserialize, Serialize, Debug, Default, Clone, Copy)]
-#[serde(rename_all = "camelCase")]
-struct JBSize {
-    width: u32,
-    height: u32,
 }
 
 #[derive(Deserialize, Debug)]
